Combine Navigation/index.js into App.js
Create placeholder custom buttons & input
