+++ Components/Styled Components/buttons.js
@@ -2,7 +2,8 @@ import { Pressable } from "react-native";
 import styled from "styled-components/native";
 
 const Button = styled(Pressable)`
-	
+	color: white;
+    background-color: blue;
 `;
 
 export { Button };
